Update ollama model name fetching mechanism.
There is a bug, that I have not been able to reproduce in the tests,
where sometimes one of the llm model names comes back as some part of
the alphanumeric ID. This rekajiggers the mechanism to get those. Still
passes tests. Hoping this accidentally takes care of it.

The main change is that it's matching against one space, rather than two
or more spaces.