Take another approach at fetching available ollama models
This pesky bug where sometimes the names are showing up as the IDs, or
some other part of what's returned, persists.

This takes a different approach - instead of string munging a response
and self implementing awk, the plugin is now fetching the types via curl
and getting a json response, and disassembling that.