Make major update to error handling
This change brings error handling back into the windows themselves,
rather than using the util.log function. This is more transparent and
helpful. This knocks out:

* Gracefully handle when ollama isn't running

  Now there's a curl error in the window indicating the server can't be
  reached.

* Show all curl errors in user window. Sad path here is unhelpful at best.

  Sad path is now less unhelpful and more not unhelpful

* Logging the errors instead of showing them in the window was the wrong
  way, I just want to handle as many as possible in the provider.
  * Provider errors might be alright to just log, actually, making sure
    all known ones are handled..

  They're now showing in the window again, _and_ all the ollama errors
  I remember seeing are handled.

There is one caveat to this change - if there is an unhandled provider
error, it will appear in the screen. I'm a little worried about openai
errors, because I have written that they sometimes split single json
responses into multiple response frames, which will appear real noisy
and buggy. But I think it's the only way to not sweep errors under the
rug.

There's a world in which the provider accumulates frames when one isn't
json decodable, but that world is not here today.