Fix bug where clipped messages displayed an error to user
I'm not sure if it was just me or what, but it seems like more and more
I was seeing json decode errors in the windows. This is desired
behavior, but there shouldn't be that many decode errors. Turns out the
servers were often splitting json responses between two response chunks,
despite curl getting the --no-buffer directive.

This change fixes that, allowing the providers to join clipped
responses, both for openai and for ollama.

It also introduces a shared test suite, which I'd love to be more filled
out.