Feature: Include LSP diagnostics in code window when extent in selected text
I at first wanted to add an entry into the code diagnostic window, but I
wasn't sure how to select the "relevant" blocks of code. It occurred to
me it'd be easy to just include it. I've been using it for a bit now and
really liking it.

I do worry it's a bit intrusive for anybody who may already be using
this, who maybe doesn't pay as much attention to diagnostics. But the
goal is for it to be simple - if you aren't trying to fix the
diagnostics, you can just <S-v>Gs and you're writing a fresh prompt.