Fix issue where ctrl-j/k froze when selected model wasn't present
Ex. if a user wasn't running the ollama server and they pressed c-j or
c-k. Since no ollama models were present, that would freeze the
scrolling.

This also allowed a bunch of selection logic to be removed, and TODOs
for more as well. Turns out I was coding around that one bug for a long
time. It wasn't even a hard fix, wild.