Move dependency checking to vim.notify_once
This fixes* a bug where every time a user opened a code or chat window
and had a saved session, the saved session gets overwritten by the error
of there being missing dependencies.

So for example a user who doesn't use ollama could never see their saved
code sessions, always seeing the missing deps error message instead.

Now it uses neovim's notify_once command, which is just the perfect
thing for this circumstance. Was so stoked to see that.

*the bug is probably still there but in the form of fetches to get
ollama models doing the same thing as the dep check.

Addresses #4