Reduce the intensity of missing deps notifications
Both missing requirements and optional deps are one line now, and the
optional deps has been downgraded to info. One user indicated the
message was too much, and thinking about it - yeah it's not a warning to
have a missing optional dependency. It's *optional*. Now the message
hardly grabs the eye at all, but it is present in the bottom, replacing
seeing :GPTModels* with the deps message. Better experience.