Fix exp for users not having openai set up; fetch all available openai models
Now there's a notification for missing the OPENAI_API_KEY on the first
launch, and all openai models are removed. Furthermore, all available
openai chat models are now available to use. It's a little noisier, but
it seems more correct. Maybe it doesn't feel as homey, and the ctrl-j/k
is less useful. But this plugin is called GPTModels, it's all about
using different models. And all the ollama models are available, why not
all the openai ones too? And all of the ones from other providers if
they ever come around?

Addresses #1