Fix: openai models not appearing
Turns out openai just added many new models, and now curl chunks the
output, which has to be manually handled since I'm not using an http
library here. This commit fixes that in a quasi-satisfactory way, in
that now the openai models are replaced every time the modal opens. This
stinks and should be fixed, and will when I have time.

Closes #7