Fix flaky "calls onread with errors" test
The `onerror` function got called multiple times, and we only stored the
error of the last call.

When the test was passing, it was called with these `err` values:
1. "ls:"
2. "cannot access '/thisdirdefinitelydoesntexistimsureofitforreal': No such file or directory"

When the test was failing, it was called with these `err` values:
1. "ls:"
2. "cannot access '/thisdirdefinitelydoesntexistimsureofitforreal'"
3. ": No such file or directory"
4. "" (empty string, space or new line?)

The solution is to store all the errors, not only the last.